requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/dev.in
#
appdirs==1.4.4
    # via black
black==21.7b0
    # via -r requirements/dev.in
certifi==2020.12.5
    # via
    #   -r requirements/test.txt
    #   requests
chardet==4.0.0
    # via
    #   -r requirements/test.txt
    #   requests
click==8.0.1
    # via
    #   -r requirements/test.txt
    #   black
    #   pip-tools
docker==5.0.0
    # via -r requirements/test.txt
idna==2.10
    # via
    #   -r requirements/test.txt
    #   requests
mypy-extensions==0.4.3
    # via black
pathspec==0.9.0
    # via black
pep517==0.10.0
    # via pip-tools
pip-tools==6.1.0
    # via -r requirements/dev.in
python-dotenv==0.17.0
    # via -r requirements/test.txt
pyyaml==5.4.1
    # via -r requirements/test.txt
regex==2021.7.6
    # via black
requests==2.25.1
    # via
    #   -r requirements/test.txt
    #   docker
six==1.15.0
    # via
    #   -r requirements/test.txt
    #   websocket-client
toml==0.10.2
    # via pep517
tomli==1.0.4
    # via black
urllib3==1.26.4
    # via
    #   -r requirements/test.txt
    #   requests
websocket-client==0.58.0
    # via
    #   -r requirements/test.txt
    #   docker

# The following packages are considered to be unsafe in a requirements file:
# pip
